requirements.txt
autopep8
mypy
pylint
pytest
python-dotenv
requests
types-requests
ics
pytz
python-dateutil
